django/requirements.txt
Django
mysqlclient
djangorestframework
